sidecar-testing/fastapi/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
-e ../../sdks/python
